+++ src/replay/meson.build
@@ -1,0 +1,10 @@
+
+replay_src = files(
+    'play.c',
+    'read.c',
+    'replay.c',
+    'rw_common.c',
+    'stage.c',
+    'state.c',
+    'write.c',
+)
